requirements.txt
flask
flask_sqlalchemy
gunicorn
psycopg2
paython-dotenv